Report For Project
In this project we are going to introduce a new way to create new features in a dataset
for machine learning algorithms. We will find interaction between features using
differential evolution . We have applied the differential evolution in its crude form which
give us a lot of future scope for improvement also.